split gesture service into more files
