+++ test/full_coverage_test.dart
@@ -42,9 +42,9 @@ import 'package:flutter_map/src/map/controller/events/map_event_source.dart';
 import 'package:flutter_map/src/map/controller/events/map_events.dart';
 import 'package:flutter_map/src/map/controller/map_controller.dart';
 import 'package:flutter_map/src/map/controller/map_controller_impl.dart';
-import 'package:flutter_map/src/map/gestures/gestures.dart';
 import 'package:flutter_map/src/map/gestures/latlng_tween.dart';
 import 'package:flutter_map/src/map/gestures/map_interactive_viewer.dart';
+import 'package:flutter_map/src/map/gestures/services/base_services.dart';
 import 'package:flutter_map/src/map/inherited_model.dart';
 import 'package:flutter_map/src/map/options/enabled_gestures.dart';
 import 'package:flutter_map/src/map/options/interaction_options.dart';
